requirements.txt
streamlit
pypdf
google-genai
python-dotenv
fpdf2
markdown
supabase
bcrypt
genanki
gTTS
langchain-text-splitters
chromadb
requests
